- Move global-question-actions → prisma/actions/global-questions.ts (R1-M1) - Move global-questions data query → prisma/data/global-questions.ts (R1-M1) - Move shared constants/base schema → lib/constants.ts (R1-M2) - Move GlobalQuestionListItem type → lib/types.ts (R1-M2) - Remove out-of-scope reorder action and UI buttons (R1-M3) - Replace no-dep useEffect ref sync with useLayoutEffect (R1-M4) - Use ES shorthand property for options in Prisma data objects (R1-N1) - Add comment explaining revalidatePath('/profile') dependency (R1-L2)
