+++ lib/types.ts
@@ -5,6 +5,7 @@ import type {
   PositionApplicationAnswer,
   PositionQuestion,
 } from '@/prisma/client';
+import type { Prisma } from '@/prisma/client';
 
 export type PositionWithQuestions = Position & {
   questions: PositionQuestion[];
@@ -14,3 +15,16 @@ export type DraftApplication = Application & {
   globalAnswers: GlobalApplicationAnswer[];
   positionAnswers: PositionApplicationAnswer[];
 };
+
+export type GlobalQuestionListItem = Prisma.GlobalQuestionGetPayload<{
+  select: {
+    id: true;
+    order: true;
+    label: true;
+    type: true;
+    required: true;
+    options: true;
+    createdAt: true;
+    updatedAt: true;
+  };
+}>;
